fix: castear unidades a FLOAT antes de dividir para evitar división entera en Redshift
La división integer/integer en Redshift trunca a 0, causando que todos
los cambio_pct_cantidad fueran 0 y por ende todas las elasticidades 0.

+++ mcdonalds_simulator/core/elasticidades.py
@@ -98,7 +98,7 @@ def calcular_y_guardar():
             SELECT
                 codigo,
                 periodo,
-                (unidades - unidades_anterior) / NULLIF(unidades_anterior, 0) AS cambio_pct_cantidad
+                (unidades - unidades_anterior)::FLOAT / NULLIF(unidades_anterior, 0) AS cambio_pct_cantidad
             FROM ventas_con_lag
             WHERE unidades_anterior IS NOT NULL
               AND unidades_anterior != unidades
